[SERVDEV-345] - Change namespace from com.pentaho to org.pentaho - Updated dependencies
