+++ pdi-det-assembly/pom.xml
@@ -18,38 +18,20 @@
   <url>http://www.pentaho.com/</url>
 
   <dependencies>
-    <!--
     <dependency>
       <groupId>pentaho</groupId>
       <artifactId>pdi-det</artifactId>
       <version>${project.version}</version>
     </dependency>
-    -->
 
-    <dependency>
-      <groupId>pentaho</groupId>
-      <artifactId>pentaho-det-core</artifactId>
-      <version>${project.version}</version>
-    </dependency>
   </dependencies>
 
   <build>
     <plugins>
 
-      <!--
       <plugin>
         <groupId>org.apache.karaf.tooling</groupId>
         <artifactId>karaf-maven-plugin</artifactId>
-      </plugin>
--->
-      <plugin>
-        <groupId>org.apache.karaf.tooling</groupId>
-        <artifactId>karaf-maven-plugin</artifactId>
-        <version>3.0.3</version>
-        <extensions>true</extensions>
-        <configuration>
-          <aggregateFeatures>false</aggregateFeatures>
-        </configuration>
       </plugin>
 
     </plugins>
