+++ pdi-det/pom.xml
@@ -38,6 +38,14 @@
       <version>${project.version}</version>
     </dependency>
 
+    <!-- Client Side dependencies -->
+    <dependency>
+      <groupId>pentaho</groupId>
+      <artifactId>common-ui</artifactId>
+      <version>${project.version}</version>
+      <type>zip</type>
+    </dependency>
+
     <!-- Test dependencies -->
     <dependency>
       <groupId>junit</groupId>
