+++ pdi-det/src/main/resources/OSGI-INF/blueprint/pdi-det-blueprint.xml
@@ -26,7 +26,7 @@
   </bean>
 
   <service id="previewDataSourceProviderService"
-           interface="com.pentaho.det.api.services.IDataSourceProvider"
+           interface="org.pentaho.det.api.services.IDataSourceProvider"
            ref="previewDataSourceProviderBean">
   </service>
