+++ pentaho-det-core/src/main/resources/web/index.html
@@ -44,7 +44,7 @@
 
     <script type="text/javascript">
       require.config( requireCfg );
-      require([ "com.pentaho.det/js/app" ], function( app ) {
+      require([ "pentaho/det/js/app" ], function( app ) {
         app.init( document );
       });
     </script>
