+++ pentaho-det-core/src/main/resources/web/js/app.html
@@ -1,5 +1,5 @@
 <div class="pentaho.det.app">
-  <!-- div data-ng-controller="com.pentaho.det.ApplicationController as appViewModel" -->
+  <!-- div data-ng-controller="org.pentaho.det.ApplicationController as appViewModel" -->
 
     <div class="navbar">
       <div class="navbar-inner">
