+++ pentaho-det-embed-examples/pom.xml
@@ -3,7 +3,7 @@
   <modelVersion>4.0.0</modelVersion>
 
   <parent>
-    <groupId>com.pentaho</groupId>
+    <groupId>pentaho</groupId>
     <artifactId>pentaho-det-bundles</artifactId>
     <version>6.0-SNAPSHOT</version>
   </parent>
