+++ pentaho-det-embed-examples/src/main/resources/web/examples/AngularWebbAppFullDET/index.html
@@ -38,7 +38,7 @@
       require.config( requireCfg );
 
 
-      require([ "com.pentaho.det/js/app" ], 
+      require([ "pentaho/det/js/app" ],
         function( app ) {
           app.init( document );
       });
