+++ pentaho-det-embed-examples/src/main/resources/web/examples/AngularWebbAppTwoModules/index.html
@@ -46,8 +46,8 @@
       require([
         "angular",
 
-        "com.pentaho.det.pluginA/app.module",
-        "com.pentaho.det.pluginB/app.module",
+        "pentaho/det/pluginA/app.module",
+        "pentaho/det/pluginB/app.module",
 
         "underscorejs",
