+++ pentaho-det-embed-examples/src/main/resources/web/examples/NotAngularWebbAppFullDET/index.html
@@ -25,7 +25,7 @@
 
       require.config( requireCfg );
 
-      require([ "com.pentaho.det/js/app" ], function( app ) {
+      require([ "pentaho/det/js/app" ], function( app ) {
         app.init( document );
       });
     </script>
