+++ pentaho-det-embed-examples/src/main/resources/web/examples/NotAngularWebbAppTwoModules/index.html
@@ -28,8 +28,8 @@
       require([ 
         "angular",
 
-        "com.pentaho.det.pluginA/app.module",
-        "com.pentaho.det.pluginB/app.module",
+        "pentaho/det/pluginA/app.module",
+        "pentaho/det/pluginB/app.module",
 
         "underscorejs",
