+++ pentaho-pluginA/pom.xml
@@ -47,7 +47,7 @@
             <Bundle-SymbolicName>${project.artifactId}</Bundle-SymbolicName>
             <Bundle-Version>${project.version}</Bundle-Version>
             <Export-Package>
-              com.pentaho.det.*
+              org.pentaho.det.*
             </Export-Package>
             <Import-Package>
               org.pentaho.di.osgi, *
