+++ pentaho-pluginA/src/main/resources/META-INF/js/require.json
@@ -1,11 +1,11 @@
 {
   "paths": {
-    "com.pentaho.det.pluginA": "DataExplorerToolPluginA/web"
+    "pentaho/det/pluginA": "DataExplorerToolPluginA/web"
   },
 
   "config": {
     "service": {
-      "com.pentaho.det.pluginA/app.module": "IDetPlugin"
+      "pentaho/det/pluginA/app.module": "IDetPlugin"
     }
   }
 }
