+++ pentaho-pluginB/src/main/resources/META-INF/js/require.json
@@ -1,11 +1,11 @@
 {
   "paths": {
-    "com.pentaho.det.pluginB": "DataExplorerToolPluginB/web"
+    "pentaho/det/pluginB": "DataExplorerToolPluginB/web"
   },
 
   "config": {
     "service": {
-      "com.pentaho.det.pluginB/app.module": "IDetPlugin"
+      "pentaho/det/pluginB/app.module": "IDetPlugin"
     }
   }
 }
