+++ pom.xml
@@ -41,11 +41,13 @@
         <groupId>pentaho-kettle</groupId>
         <artifactId>kettle-core</artifactId>
         <version>${dependency.pentaho-kettle.version}</version>
+        <scope>provided</scope>
       </dependency>
       <dependency>
         <groupId>pentaho-kettle</groupId>
         <artifactId>kettle-ui-swt</artifactId>
         <version>${dependency.pentaho-kettle.version}</version>
+        <scope>provided</scope>
       </dependency>
 
       <dependency>
@@ -68,6 +70,7 @@
         <groupId>javax.ws.rs</groupId>
         <artifactId>javax.ws.rs-api</artifactId>
         <version>2.0</version>
+        <scope>provided</scope>
       </dependency>
 
       <dependency>
@@ -117,6 +120,24 @@
         <version>2.1</version>
       </dependency>
 
+      <!-- TODO verify jackson serializer -->
+      <dependency>
+        <groupId>org.codehaus.jackson</groupId>
+        <artifactId>jackson-jaxrs</artifactId>
+        <version>1.9.13</version>
+      </dependency>
+
+      <dependency>
+        <groupId>org.codehaus.jackson</groupId>
+        <artifactId>jackson-core-asl</artifactId>
+        <version>1.9.13</version>
+      </dependency>
+
+      <dependency>
+        <groupId>org.codehaus.jackson</groupId>
+        <artifactId>jackson-mapper-asl</artifactId>
+        <version>1.9.13</version>
+      </dependency>
 
       <!-- Test Dependencies -->
       <dependency>
